requirements.txt
tensorflow==2.15.0
opendatasets==0.1.22
Pillow==10.2.0
numpy==1.24.3
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
tf-keras-vis==0.8.6
opencv-python==4.9.0.80 